+++ crucible-llvm-syntax/CHANGELOG.md
@@ -1,6 +1,7 @@
 # next
 
-Nothing yet.
+* Add overrides for string-manipulation: `{read,write}-bytes` and
+  `{read,write}-c-string`.
 
 # 0.1 -- 2024-02-05